Separated views and API
